1_foundations/requirements.txt
requests
python-dotenv
gradio
pypdf2
openai
openai-agents